refactor(router): move applyServerActionResultDecision to dedicated navigation module
